move some things to e/toolkit
