+++ package.json
@@ -22,7 +22,7 @@
 		"@fastify/sensible": "^5.2.0",
 		"@napi-rs/canvas": "^0.1.38",
 		"@octokit/graphql": "^4.8.0",
-		"@oldschoolgg/toolkit": "^0.0.12",
+		"@oldschoolgg/toolkit": "^0.0.13",
 		"@prisma/client": "^3.15.1",
 		"@sapphire/stopwatch": "^1.4.0",
 		"@sapphire/time-utilities": "^1.6.0",
@@ -35,14 +35,12 @@
 		"discord.js": "^14.8.0",
 		"dotenv": "^16.0.3",
 		"dotenv-cli": "^7.1.0",
-		"e": "^0.2.32",
-		"emoji-regex": "^10.2.1",
+		"e": "^0.2.33",
 		"fastify": "^4.14.1",
 		"fastify-raw-body": "^4.2.0",
 		"he": "^1.2.0",
 		"lru-cache": "^8.0.0",
 		"mahoji": "^0.0.62",
-		"math-expression-evaluator": "^1.3.14",
 		"murmurhash": "^2.0.1",
 		"node-cron": "^3.0.2",
 		"node-fetch": "^2.6.7",
@@ -62,7 +60,6 @@
 		"@oldschoolgg/ts-config": "^0.0.1",
 		"@types/he": "^1.1.2",
 		"@types/madge": "^5.0.0",
-		"@types/math-expression-evaluator": "^1.2.2",
 		"@types/node": "^14.18.12",
 		"@types/node-cron": "^3.0.7",
 		"@types/node-fetch": "^2.6.1",
