+++ src/lib/data/Collections.ts
@@ -1,5 +1,5 @@
 import { AttachmentBuilder } from 'discord.js';
-import { calcWhatPercent, isObject, notEmpty, sumArr, uniqueArr } from 'e';
+import { calcWhatPercent, isObject, notEmpty, removeFromArr, sumArr, uniqueArr } from 'e';
 import { Bank, Clues, Monsters } from 'oldschooljs';
 import { Item } from 'oldschooljs/dist/meta/types';
 import { ChambersOfXeric } from 'oldschooljs/dist/simulation/misc/ChambersOfXeric';
@@ -22,10 +22,9 @@ import { getToaKCs } from '../simulation/toa';
 import { allFarmingItems } from '../skilling/skills/farming';
 import { SkillsEnum } from '../skilling/types';
 import type { ItemBank } from '../types';
-import { fetchStatsForCL } from '../util';
-import { stringMatches } from '../util/cleanString';
+import { fetchStatsForCL, stringMatches } from '../util';
 import resolveItems from '../util/resolveItems';
-import { removeFromArr, shuffleRandom } from '../util/smallUtils';
+import { shuffleRandom } from '../util/smallUtils';
 import {
 	abyssalSireCL,
 	aerialFishingCL,
