+++ src/lib/skilling/skills/index.ts
@@ -74,6 +74,4 @@ export const Skills: Record<string, Skill> = {
 	}
 };
 
-export const skillsValues = Object.values(Skills);
-
 export default Skills;
