--- src/lib/util/cleanString.ts
@@ -1,6 +1,0 @@
-export function cleanString(str: string) {
-	return str.replace(/[^0-9a-zA-Z+]/gi, '').toUpperCase();
-}
-export function stringMatches(str: string | number = '', str2: string | number = '') {
-	return cleanString(str.toString()) === cleanString(str2.toString());
-}
