+++ src/lib/util/getKCByName.ts
@@ -1,9 +1,9 @@
+import { stringMatches } from '@oldschoolgg/toolkit';
 import Monster from 'oldschooljs/dist/structures/Monster';
 
 import { effectiveMonsters } from '../minions/data/killableMonsters';
 import { getMinigameScore, Minigames } from '../settings/minigames';
 import creatures from '../skilling/skills/hunter/creatures';
-import { stringMatches } from './cleanString';
 
 export async function getKCByName(user: MUser, kcName: string): Promise<[string, number] | [null, 0]> {
 	const mon = effectiveMonsters.find(
