+++ src/lib/util/minionStatus.ts
@@ -72,8 +72,7 @@ import {
 	WoodcuttingActivityTaskOptions,
 	ZalcanoActivityTaskOptions
 } from '../types/minions';
-import { formatDuration, itemNameFromID, randomVariation } from '../util';
-import { stringMatches } from './cleanString';
+import { formatDuration, itemNameFromID, randomVariation, stringMatches } from '../util';
 import { getActivityOfUser } from './minionIsBusy';
 import { toTitleCase } from './toTitleCase';
