+++ src/lib/util/parseStringBank.ts
@@ -1,10 +1,10 @@
+import { evalMathExpression } from '@oldschoolgg/toolkit/dist/util/expressionParser';
 import { notEmpty } from 'e';
 import { Bank, Items } from 'oldschooljs';
 import { Item } from 'oldschooljs/dist/meta/types';
 import { itemNameMap } from 'oldschooljs/dist/structures/Items';
 
 import { filterableTypes } from '../data/filterables';
-import { evalMathExpression } from '../expressionParser';
 import { cleanString, stringMatches } from '../util';
 import itemIsTradeable from './itemIsTradeable';
