+++ src/mahoji/lib/abstracted_commands/autoSlayCommand.ts
@@ -1,3 +1,4 @@
+import { isGuildChannel } from '@oldschoolgg/toolkit';
 import { ChatInputCommandInteraction } from 'discord.js';
 import { CommandOptions } from 'mahoji/dist/lib/types';
 import { Monsters } from 'oldschooljs';
@@ -7,7 +8,7 @@ import killableMonsters from '../../../lib/minions/data/killableMonsters';
 import { runCommand } from '../../../lib/settings/settings';
 import { autoslayModes, AutoslayOptionsEnum } from '../../../lib/slayer/constants';
 import { getCommonTaskName, getUsersCurrentSlayerInfo, SlayerMasterEnum } from '../../../lib/slayer/slayerUtil';
-import { hasSkillReqs, isGuildChannel, stringMatches } from '../../../lib/util';
+import { hasSkillReqs, stringMatches } from '../../../lib/util';
 import { slayerNewTaskCommand } from './slayerTaskCommand';
 
 interface AutoslayLink {
